LayoutTests/platform/mac-elcapitan/media/track/video-track-alternate-groups-expected.txt


Video elements containing audio tracks with alternate groups.


Test that media containing three audio tracks with no alternate group results in three AudioTracks
RUN(video.src = findMediaFile("video", "../content/audio-tracks-no-alternate-group"))
EVENT(canplaythrough)
EXPECTED (video.audioTracks.length == '3') OK
TRACK id:2 kind:main language:eng
TRACK id:3 kind:main language:fra
TRACK id:4 kind:main language:spa

Test that media containing two audio tracks in one alternate group and one track without an alternate group results in two AudioTracks
RUN(video.src = findMediaFile("video", "../content/audio-tracks-some-in-alternate-group"))
EVENT(canplaythrough)
EXPECTED (video.audioTracks.length == '2') OK
TRACK id:2 kind:main language:en
TRACK id:3 kind:main language:fr

Test that media containing six audio tracks, with three audio tracks each referring to a separate track as a fallback track, all in one alternate group, results in three AudioTracks
RUN(video.src = findMediaFile("video", "../content/audio-tracks-alternate-group-with-fallback"))
EVENT(canplaythrough)
EXPECTED (video.audioTracks.length == '3') OK
TRACK id:3 kind:main language:en
TRACK id:5 kind:main language:fr
TRACK id:7 kind:main language:es
END OF TEST
